Now calculates time to enable powered flight including air resistance
Time for powered descent is way later if air resistance is included
So we simulate with air resistance then found the last point where the
required suicide velocity crosses the actual velocity at that height
If the vessel slows enough the burn can be late. This is re-calculated
every time the controller tests moving into powered descent. So the
altitude to enable to lowered