More tweak of gains
With more accurate calculation of aero forces, still factor 2x3 off I think
I can now use the calculated effect of steer angle to alter gain rather than just the sign
This means less steering with very sensitive craft like in FAR with grid fins
It should mean that the gain doesn't need to be reduced so much
I also add a large deadzone where Fthrust and Faero and within a factor
of each other, the aeroThrustPropMargin in which to do no steering as
direction might be wrong. Safer to not steer
I tweak aero force up by factor of 2? as this seemed to be low